카테고리명 출력, 관리자 후원승인 기능 구현 필요

+++ src/main/java/mapper/CategoryMapper.java
@@ -7,8 +7,8 @@ import domain.Category;
 
 public interface CategoryMapper {
 	void insert(Category cate);
-	void delete(Long cno);
+	void delete(Integer cno);
 	void update(Category cate);
 	
-	Category selectOne(Long cno);
+	Category selectOne(Integer cno);
 }
